requirements.txt
accelerate==1.7.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
audioread==3.0.1
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
chatterbox-tts==0.1.1
click==8.2.1
conformer==0.3.2
decorator==5.2.1
diffusers==0.29.0
distlib==0.3.9
einops==0.8.1
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.2
identify==2.6.12
idna==3.10
importlib-metadata==8.7.0
jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.62
langchain-ollama==0.3.3
langchain-text-splitters==0.3.8
langsmith==0.3.43
lazy-loader==0.4
librosa==0.10.0
llvmlite==0.44.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.5
nltk==3.9.1
nodeenv==1.9.1
numba==0.61.2
numpy==1.26.0
ollama==0.4.9
omegaconf==2.3.0
onnx==1.18.0
openai-whisper==20240930
orjson==3.10.18
packaging==24.2
peft==0.15.2
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
pre-commit==4.2.0
protobuf==6.31.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.5
pydantic-core==2.33.2
pygments==2.19.1
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
resampy==0.4.3
resemble-perth==1.0.1
rich==14.0.0
s3tokenizer==0.1.7
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.9.0
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
soxr==0.5.0.post1
sqlalchemy==2.0.41
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.20.3
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
transformers==4.46.3
typing-extensions==4.13.2
typing-inspection==0.4.1
urllib3==2.4.0
virtualenv==20.31.2
zipp==3.22.0
zstandard==0.23.0

# Piper TTS (fast local TTS)
piper-tts==1.3.0
onnxruntime==1.18.1

# Google Gemini API
google-generativeai==0.8.3

# Memory system dependencies
psycopg2-binary==2.9.9
sentence-transformers==2.7.0
